requirements.txt
# This file is managed by openstack-depends
cliff>=1.4
d2to1>=0.2.10,<0.3
eventlet>=0.9.12
extras
flask==0.9
ipaddr
iso8601>=0.1.4
jsonschema>=1.0.0,!=1.4.0,<2
kombu>2.4.7
oslo.config>=1.1.0
paste
pastedeploy>=1.5.0
pbr>=0.5.16,<0.6
python-keystoneclient>=0.2,<0.3
routes>=1.12.3
sqlalchemy>=0.7,<=0.7.99
sqlalchemy-migrate>=0.7
stevedore>=0.9
webob>=1.0.8
